+++ backend/ddang/src/test/java/com/ddang/ddang/report/presentation/ReportControllerTest.java
@@ -5,7 +5,7 @@ import com.ddang.ddang.authentication.configuration.AuthenticationInterceptor;
 import com.ddang.ddang.authentication.configuration.AuthenticationPrincipalArgumentResolver;
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
-import com.ddang.ddang.authentication.domain.dto.AuthenticationStore;
+import com.ddang.ddang.authentication.configuration.AuthenticationStore;
 import com.ddang.ddang.chat.infrastructure.exception.ChatRoomNotFoundException;
 import com.ddang.ddang.exception.GlobalExceptionHandler;
 import com.ddang.ddang.qna.infrastructure.exception.AnswerNotFoundException;
