+++ backend/ddang/src/test/java/com/ddang/ddang/review/presentation/ReviewControllerTest.java
@@ -24,7 +24,7 @@ import com.ddang.ddang.authentication.configuration.AuthenticationInterceptor;
 import com.ddang.ddang.authentication.configuration.AuthenticationPrincipalArgumentResolver;
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
-import com.ddang.ddang.authentication.domain.dto.AuthenticationStore;
+import com.ddang.ddang.authentication.configuration.AuthenticationStore;
 import com.ddang.ddang.exception.GlobalExceptionHandler;
 import com.ddang.ddang.review.application.dto.request.CreateReviewDto;
 import com.ddang.ddang.review.application.exception.AlreadyReviewException;
