+++ backend/ddang/src/test/java/com/ddang/ddang/user/presentation/UserAuctionControllerTest.java
@@ -23,7 +23,7 @@ import com.ddang.ddang.authentication.configuration.AuthenticationInterceptor;
 import com.ddang.ddang.authentication.configuration.AuthenticationPrincipalArgumentResolver;
 import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
-import com.ddang.ddang.authentication.domain.dto.AuthenticationStore;
+import com.ddang.ddang.authentication.configuration.AuthenticationStore;
 import com.ddang.ddang.exception.GlobalExceptionHandler;
 import com.ddang.ddang.user.presentation.fixture.UserAuctionControllerFixture;
 import java.util.Optional;
